src/requirements.txt
fastapi
torch==2.9.0
uvicorn[standard]
pydantic
pymongo
sqlalchemy
pandas
python-dotenv
unsloth[colab-new]
transformers
accelerate
bitsandbytes
mysql-connector-python
tabulate